#32 Move InlineAlert into main Alert directory
