+++ src/components/Alert/InlineAlert.test.tsx
@@ -1,24 +1,22 @@
 import { mount } from 'enzyme';
 import * as React from 'react';
-import * as Rivet from '../Rivet';
-import Input from '../Input/Input';
 import InlineAlert from './InlineAlert';
 
 describe('Inline Alerts', () => {
     it('should have "info" styling with the info variant', () => { 
-        const cut = mount(<InlineAlert variant="info" >🤔</InlineAlert>);
+        const cut = mount(<InlineAlert variant="info">🤔</InlineAlert>);
         expect(cut.find('.rvt-inline-alert').hasClass("rvt-inline-alert--has-info")).toBe(true);
     });
     it('should have "valid" styling with the valid variant', () => { 
-        const cut = mount(<InlineAlert variant="valid" >😎</InlineAlert>);
+        const cut = mount(<InlineAlert variant="valid">😎</InlineAlert>);
         expect(cut.find('.rvt-inline-alert').hasClass("rvt-inline-alert--is-valid")).toBe(true);
     });
     it('should have "warning" styling with the warning variant', () => { 
-        const cut = mount(<InlineAlert variant="warning" >🤨</InlineAlert>);
+        const cut = mount(<InlineAlert variant="warning">🤨</InlineAlert>);
         expect(cut.find('.rvt-inline-alert').hasClass("rvt-inline-alert--has-warning")).toBe(true);
     });
     it('should have "invalid" styling with the invalid variant', () => { 
-        const cut = mount(<InlineAlert variant="invalid" >😬</InlineAlert>);
+        const cut = mount(<InlineAlert variant="invalid">😬</InlineAlert>);
         expect(cut.find('.rvt-inline-alert').hasClass("rvt-inline-alert--is-invalid")).toBe(true);
     });
     it('applies class names to the containing div', () => {
@@ -28,9 +26,9 @@ describe('Inline Alerts', () => {
 
     it('renders the children as a message of the alert', () => {
         const messageText = 'This is some alert text'
-        const cut = mount(<InlineAlert variant="info" >{messageText}</InlineAlert>);
+        const cut = mount(<InlineAlert variant="info">{messageText}</InlineAlert>);
         expect(cut.find('.rvt-inline-alert').text()).toBe(messageText);
-    })
+    });
 });
 
 describe('Standalone Inline Alerts', () => {
