--- src/components/Alert/InlineAlert/index.ts
@@ -1,1 +1,0 @@
-export { default as InlineAlert } from './InlineAlert';
