+++ src/components/Alert/index.ts
@@ -1,2 +1,3 @@
 export { default as Alert } from './Alert';
 export { default as DismissibleAlert } from './DismissibleAlert';
+export { default as InlineAlert } from './InlineAlert';
